+++ examples/sharded_array_write_read.rs
@@ -14,7 +14,7 @@ fn sharded_array_write_read() -> Result<(), Box<dyn std::error::Error>> {
     use std::sync::Arc;
 
     // Create a store
-    // let path = tempdir::TempDir::new("example")?;
+    // let path = tempfile::TempDir::new()?;
     // let store = Arc::new(store::filesystem::FilesystemStore::new(path.path())?);
     let store = Arc::new(store::MemoryStore::default());